+++ deepinv/training/trainer.py
@@ -626,11 +626,9 @@ class Trainer:
                     meters.update(loss.detach().cpu().numpy())
                     logs[l.__class__.__name__] = meters.avg
 
-                meters = (
-                    self.logs_total_loss_train if train else self.logs_total_loss_eval
-                )
-                meters.update(loss_total.item())
-                logs[f"TotalLoss"] = meters.avg
+            meters = self.logs_total_loss_train if train else self.logs_total_loss_eval
+            meters.update(loss_total.item())
+            logs[f"TotalLoss"] = meters.avg
         else:  # TODO question: what do we want to do at test time?
             loss_total = 0
 
@@ -940,6 +938,9 @@ class Trainer:
 
         for l in self.logs_metrics_eval:
             l.reset()
+
+        if hasattr(self, "check_grad_val"):
+            self.check_grad_val.reset()
 
     def save_best_model(self, epoch, train_ite, **kwargs):
         r"""